requirements.txt
asgiref
Django
djangorestframework
pillow
psycopg2-binary
sqlparse
geojson
django-import-export
langdetect
